Don't unify synthetic types between method parameters.
Currently, if two parameters (and/or the return type) share the same structural type in javascript, the java synthetic types are unified. This logic could cause future breaking change on java side for a non-breaking change in javascript.

We are fixing the issue by fixing the way we calculate a string representation of a type structure in order to take into account any parameter position.

Elemental source diff: cl/160438706

PiperOrigin-RevId: 168790803