Replace PACKAGE_NAME with package_name() in bzl files.
LSC document:
  https://docs.google.com/document/d/19GK3_SjUZeLxiML50Hz41p6hzajKgQHW51p9M2FK3SM/edit

PACKAGE_NAME was introduced in Blaze 10 years ago. Although it looks like a constant, its value depends on which BUILD file calls it. It is also magic: you can use it inside a macro; but not in top level of a .bzl file; and not inside a custom rule. In other words, it can be used exactly in the same places as glob or the native rules. For this reason, we want to replace the ?constant? with a proper function to be consistent with glob and others.

This change will bring more consistency. It will simplify the code in Blaze (remove support for ?dynamic values?), the spec, and the documentation.

Tested:
    tap_presubmit: http://test/OCL:187613240:BASE:187851115:1520275776407:b5f554b2
    Some tests failed; test failures are believed to be unrelated to this CL
PiperOrigin-RevId: 188011233