Revive global_scope_class_name attribute in order to be able to fix the name of the class representing the global scope.
Currently the logic for calculating the name of the class representing the global scope is driven by the dependency of the lib. So adding a dependency can rename the class. This CL fixes that by hardcoding the name of class via the skylark attribute global_scope_class_name or if this attribute is not set, by setting the class name to extension_prefix + Global in order to be backward compatible.

PiperOrigin-RevId: 201426841